# swagger.yaml
openapi: 3.0.4
info:
  title: Next Cinema API Demo 1.0
  description: |-
    To be added

    Some useful links:
    - [Github Repo](https://github.com/alexmita04/next-cinema-api)
  contact:
    email: alexmita04@gmail.com
  license:
    name: MIT
    url: https://opensource.org/license/mit
  version: 1.0.0
...
servers:
  - url: https://next-cinema-api.onrender.com
...
paths:
...
  /api/cinemas:
    get:
      tags:
        - cinema
      summary: Retrieve a list of all cinemas
      description: Returns an array of all cinema objects currently in the database.
      operationId: getAllCinemas
      responses:
        "200":
          description: Successfully retrieved the list of cinemas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllCinemasResponse"
        "500":
          description: Internal Server Error - Unexpected error during retrieval.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Cinema:
      type: object
      description: A cinema object containing its details.
      properties:
        id:
          type: string
          description: The unique ID of the cinema.
          example: 60a7e6b0c7a5f60015b3c5d9
        name:
          type: string
          description: The name of the cinema.
          example: Cineplexx Baneasa
        city:
          type: string
          description: The city where the cinema is located.
          example: Bucharest
        address:
          type: string
          description: The physical address of the cinema.
          example: Şoseaua București-Ploiești 42D
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the cinema record was created.
          example: 2021-05-21T10:00:00.000Z
      required:
        - id
        - name
        - city
        - address
    GetAllCinemasResponse:
      type: object
      description: The successful response body for the getAllCinemas operation.
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            cinemas:
              type: array
              items:
                $ref: "#/components/schemas/Cinema"
              description: The array of cinema objects.
            length:
              type: integer
              description: The number of cinemas returned.
              example: 5
          required:
            - cinemas
            - length
      required:
        - status
        - data

    # --- Error Schema (Matches ExpressError) ---
    ErrorResponse:
      type: object
      description: Generic error structure returned by the API.
      properties:
        status:
          type: string
          description: The status flag (typically 'error').
          example: error
        message:
          type: string
          description: A detailed, human-readable error message.
          example: Invalid Credentials
        statusCode:
          type: integer
          description: The HTTP status code of the error.
          example: 401